requirements.txt
Django==1.6.1
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
South==0.8.4
Werkzeug==0.9.4
django-ace==1.0.0
django-filter==0.7
djangorestframework==2.3.10
gevent==1.0
greenlet==0.4.1
itsdangerous==0.23
locustio==0.6.2
phonenumbers==5.9.1
requests==2.2.1
wsgiref==0.1.2
